--- a/.gitignore
+++ b/.gitignore
@@ -71,3 +71,5 @@ docs/references/*.out
 .pytest_cache
 
 pyuvdata/uvdata/src/miriad_wrap.cpp
+
+pyuvdata/uvdata/corr_fits_src/corr_fits.c