--- a/ci/pyuvdata_min_deps_tests.yml
+++ b/ci/pyuvdata_min_deps_tests.yml
@@ -9,6 +9,7 @@ dependencies:
   - h5py
   - coverage
   - pytest-cov
+  - cython
   - setuptools_scm
   - pip
   - pip: