--- a/ci/pyuvdata_tests.yml
+++ b/ci/pyuvdata_tests.yml
@@ -12,6 +12,7 @@ dependencies:
   - scipy
   - coverage
   - pytest-cov
+  - cython
   - setuptools_scm
   - pip
   - pip: