--- a/scripts/test_cover_noinstall.sh
+++ b/scripts/test_cover_noinstall.sh
@@ -3,7 +3,7 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd $DIR/..
 
-cd pyuvdata/tests
-python -m pytest --cov=pyuvdata --cov-config=../../.coveragerc\
-       --cov-report term --cov-report html:cover \
+cd pyuvdata
+python -m pytest --cov=pyuvdata --cov-config=../.coveragerc\
+       --cov-report term --cov-report html:tests/cover \
        "$@"