--- a/scripts/test_coverage.sh
+++ b/scripts/test_coverage.sh
@@ -5,7 +5,7 @@ cd $DIR/..
 
 python setup.py install
 
-cd pyuvdata/tests
-python -m pytest --cov=pyuvdata --cov-config=../../.coveragerc\
-       --cov-report term --cov-report html:cover \
+cd pyuvdata
+python -m pytest --cov=pyuvdata --cov-config=../.coveragerc\
+       --cov-report term --cov-report html:tests/cover \
        "$@"