<!--- Provide a general summary of your changes in the Title above -->
Some general vectorization and speed up of code.
## Description
<!--- Describe your changes in detail -->
Vectorized some for loop with numpy instead.
Added `_corr_fits` C extension to perform the bulk of the mwa_corr_fits reading 
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. If this PR closes an issue, put the word 'closes' before the issue link to auto-close the issue when the PR is merged. -->
Things were slow before, now they go vroom.
Further work on the way to #800

## Checklist:
<!--- You may remove the checklists that don't apply to your change type(s) or just leave them empty -->
<!--- Go over all the following points, and replace the space with an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I have read the [contribution guide](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/blob/master/.github/CONTRIBUTING.md).
- [x] My code follows the code style of this project.
- [x] I have updated the [CHANGELOG](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/blob/master/CHANGELOG.md).


I will also attach some profiling output I used to make these changes.
This first is a break out of the `test_mwa_corr_fits.py::test_multiple_coarse` test before (above) and after (below)
### **Before**
```
None
Wrote profile results to profile_corr_fits.py.lprof
Timer unit: 1e-06 s

Total time: 42.5514 s
File: ./profile_corr_fits.py
Function: read_multi at line 27

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    27                                           @profile
    28                                           def read_multi():
    29         1          3.0      3.0      0.0   order1 = [filelist[0:3]]
    30         1          1.0      1.0      0.0   order2 = [filelist[0], filelist[2], filelist[1]]
    31         1       2223.0   2223.0      0.0   mwa_uv1 = UVData()
    32         1       2069.0   2069.0      0.0   mwa_uv2 = UVData()
    33         1          1.0      1.0      0.0   messages = [
    34         1          1.0      1.0      0.0       "telescope_location is not set",
    35         1          0.0      0.0      0.0       "coarse channels are not contiguous for this observation",
    36         1          1.0      1.0      0.0       "some coarse channel files were not submitted",
    37                                            ]
    38         2   20137511.0 10068755.5     47.3   uvtest.checkWarnings(
    39         1          1.0      1.0      0.0       mwa_uv1.read, func_args=[order1], nwarnings=3, message=messages
    40                                            )
    41         2   22042822.0 11021411.0     51.8   uvtest.checkWarnings(
    42         1          1.0      1.0      0.0       mwa_uv2.read, func_args=[order2], nwarnings=3, message=messages
    43                                            )
    44         1     366779.0 366779.0      0.9   assert mwa_uv1 == mwa_uv2

Total time: 5.75414 s
File: /home/matthew/work/src/pyuvdata/pyuvdata/utils.py
Function: get_lst_for_time at line 1095

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  1095                                           @profile
  1096                                           def get_lst_for_time(jd_array, latitude, longitude, altitude):
  1097                                               """
  1098                                               Get the lsts for a set of jd times at an earth location.
  1099                                           
  1100                                               Parameters
  1101                                               ----------
  1102                                               jd_array : ndarray of float
  1103                                                   JD times to get lsts for.
  1104                                               latitude : float
  1105                                                   Latitude of location to get lst for in degrees.
  1106                                               longitude : float
  1107                                                   Longitude of location to get lst for in degrees.
  1108                                               altitude : float
  1109                                                   Altitude of location to get lst for in meters.
  1110                                           
  1111                                               Returns
  1112                                               -------
  1113                                               ndarray of float
  1114                                                   LSTs in radians corresponding to the jd_array.
  1115                                           
  1116                                               """
  1117         2       2308.0   1154.0      0.0      lst_array = np.zeros_like(jd_array)
  1118       244      35678.0    146.2      0.6      for ind, jd in enumerate(np.unique(jd_array)):
  1119       484     259382.0    535.9      4.5          t = Time(
  1120       242        214.0      0.9      0.0              jd,
  1121       242        212.0      0.9      0.0              format="jd",
  1122       242      71377.0    294.9      1.2              location=(Angle(longitude, unit="deg"), Angle(latitude, unit="deg")),
  1123                                                   )
  1124                                           
  1125                                                   # avoid errors if iers.conf.auto_max_age is set to None, as we do in
  1126                                                   # testing if the iers url is down
  1127       242      16931.0     70.0      0.3          if iers.conf.auto_max_age is None:  # pragma: no cover
  1128                                                       delta, status = t.get_delta_ut1_utc(return_status=True)
  1129                                                       if status in (iers.TIME_BEFORE_IERS_RANGE, iers.TIME_BEYOND_IERS_RANGE):
  1130                                                           warnings.warn(
  1131                                                               "time is out of IERS range, setting delta ut1 utc to "
  1132                                                               "extrapolated value"
  1133                                                           )
  1134                                                           t.delta_ut1_utc = delta
  1135                                           
  1136       484       6747.0     13.9      0.1          lst_array[
  1137       242    2069987.0   8553.7     36.0              np.where(np.isclose(jd, jd_array, atol=1e-6, rtol=1e-12))
  1138       242    3291304.0  13600.4     57.2          ] = t.sidereal_time("apparent").radian
  1139                                           
  1140         2          3.0      1.5      0.0      return lst_array

Total time: 33.3523 s
File: /home/matthew/work/src/pyuvdata/pyuvdata/uvdata/mwa_corr_fits.py
Function: read_mwa_corr_fits at line 168

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   168                                               @profile
   169                                               def read_mwa_corr_fits(
   170                                                   self,
   171                                                   filelist,
   172                                                   use_cotter_flags=False,
   173                                                   correct_cable_len=False,
   174                                                   phase_to_pointing_center=False,
   175                                                   run_check=True,
   176                                                   check_extra=True,
   177                                                   run_check_acceptability=True,
   178                                                   flag_init=True,
   179                                                   edge_width=80e3,
   180                                                   start_flag=2.0,
   181                                                   end_flag=2.0,
   182                                                   flag_dc_offset=True,
   183                                               ):
   184                                                   """
   185                                                   Read in MWA correlator gpu box files.
   186                                           
   187                                                   Parameters
   188                                                   ----------
   189                                                   filelist : list of str
   190                                                       The list of MWA correlator files to read from. Must include at
   191                                                       least one fits file and only one metafits file per data set.
   192                                                       Can also be a list of lists to read multiple data sets.
   193                                                   axis : str
   194                                                       Axis to concatenate files along. This enables fast concatenation
   195                                                       along the specified axis without the normal checking that all other
   196                                                       metadata agrees. This method does not guarantee correct resulting
   197                                                       objects. Please see the docstring for fast_concat for details.
   198                                                       Allowed values are: 'blt', 'freq', 'polarization'. Only used if
   199                                                       multiple files are passed.
   200                                                   use_cotter_flags : bool
   201                                                       Option to use cotter output mwaf flag files. Otherwise flagging
   202                                                       will only be applied to missing data and bad antennas.
   203                                                   correct_cable_len : bool
   204                                                       Option to apply a cable delay correction.
   205                                                   phase_to_pointing_center : bool
   206                                                       Option to phase to the observation pointing center.
   207                                                   run_check : bool
   208                                                       Option to check for the existence and proper shapes of parameters
   209                                                       after after reading in the file (the default is True,
   210                                                       meaning the check will be run).
   211                                                   check_extra : bool
   212                                                       Option to check optional parameters as well as required ones (the
   213                                                       default is True, meaning the optional parameters will be checked).
   214                                                   run_check_acceptability : bool
   215                                                       Option to check acceptable range of the values of parameters after
   216                                                       reading in the file (the default is True, meaning the acceptable
   217                                                       range check will be done).
   218                                                   flag_init: bool
   219                                                       Set to True in order to do routine flagging of coarse channel edges,
   220                                                       start or end integrations, or the center fine channel of each coarse
   221                                                       channel. See associated keywords.
   222                                                   edge_width: float
   223                                                       Only used if flag_init is True. The width to flag on the edge of
   224                                                       each coarse channel, in hz. Errors if not equal to integer multiple
   225                                                       of channel_width. Set to 0 for no edge flagging.
   226                                                   start_flag: float
   227                                                       Only used if flag_init is True. The number of seconds to flag at the
   228                                                       beginning of the observation. Set to 0 for no flagging. Errors if
   229                                                       not equal to an integer multiple of the integration time.
   230                                                   end_flag: floats
   231                                                       Only used if flag_init is True. The number of seconds to flag at the
   232                                                       end of the observation. Set to 0 for no flagging. Errors if not
   233                                                       equal to an integer multiple of the integration time.
   234                                                   flag_dc_offset: bool
   235                                                       Only used if flag_init is True. Set to True to flag the center fine
   236                                                       channel of each coarse channel.
   237                                           
   238                                                   Raises
   239                                                   ------
   240                                                   ValueError
   241                                                       If required files are missing or multiple files metafits files are
   242                                                       included in filelist.
   243                                                       If files from different observations are included in filelist.
   244                                                       If files in fileslist have different fine channel widths
   245                                                       If file types other than fits, metafits, and mwaf files are included
   246                                                       in filelist.
   247                                           
   248                                                   """
   249         2         10.0      5.0      0.0          metafits_file = None
   250         2          6.0      3.0      0.0          ppds_file = None
   251         2          7.0      3.5      0.0          obs_id = None
   252         2          6.0      3.0      0.0          file_dict = {}
   253         2          7.0      3.5      0.0          start_time = 0.0
   254         2          6.0      3.0      0.0          end_time = 0.0
   255         2          7.0      3.5      0.0          included_file_nums = []
   256         2          8.0      4.0      0.0          cotter_warning = False
   257         2          7.0      3.5      0.0          num_fine_chans = 0
   258                                           
   259                                                   # iterate through files and organize
   260                                                   # create a list of included coarse channels
   261                                                   # find the first and last times that have data
   262         8         32.0      4.0      0.0          for file in filelist:
   263         6         30.0      5.0      0.0              if file.lower().endswith(".metafits"):
   264                                                           # force only one metafits file
   265         2          7.0      3.5      0.0                  if metafits_file is not None:
   266                                                               raise ValueError("multiple metafits files in filelist")
   267         2          6.0      3.0      0.0                  metafits_file = file
   268         4         17.0      4.2      0.0              elif file.lower().endswith(".fits"):
   269                                                           # check if ppds file
   270         4         14.0      3.5      0.0                  try:
   271         4      16217.0   4054.2      0.0                      fits.getheader(file, extname="ppds")
   272                                                               ppds_file = file
   273         4         18.0      4.5      0.0                  except Exception:
   274                                                               # check obsid
   275         4       7899.0   1974.8      0.0                      head0 = fits.getheader(file, 0)
   276         4         22.0      5.5      0.0                      if obs_id is None:
   277         2        274.0    137.0      0.0                          obs_id = head0["OBSID"]
   278                                                               else:
   279         2        252.0    126.0      0.0                          if head0["OBSID"] != obs_id:
   280                                                                       raise ValueError(
   281                                                                           "files from different observations submitted "
   282                                                                           "in same list"
   283                                                                       )
   284                                                               # check headers for first and last times containing data
   285         4      16355.0   4088.8      0.0                      headstart = fits.getheader(file, 1)
   286         4      14373.0   3593.2      0.0                      headfin = fits.getheader(file, -1)
   287         4        964.0    241.0      0.0                      first_time = headstart["TIME"] + headstart["MILLITIM"] / 1000.0
   288         4        885.0    221.2      0.0                      last_time = headfin["TIME"] + headfin["MILLITIM"] / 1000.0
   289         4         16.0      4.0      0.0                      if start_time == 0.0:
   290         2          8.0      4.0      0.0                          start_time = first_time
   291                                                               # check that files with a timing offset can be aligned
   292         2        271.0    135.5      0.0                      elif np.abs(start_time - first_time) % headstart["INTTIME"] != 0.0:
   293                                                                   raise ValueError(
   294                                                                       "coarse channel start times are misaligned by an amount that is not \
   295                                                                           an integer multiple of the integration time"
…