@mkolopanis 2020-04-28
I think @PyxieLouStar  should have a look at what I've changed in corr_fits. She definitely understands that code better than anyone and probably can tell if I messed anything up.